Update Model and View to display the end game screen
Model has a new variable is_game_over that keeps track whether the game
is over or not. View has a new boolean variable end_drawn to make sure
that draw_end_game is only called once (so that the translucency does
not stack on top of itself). View also sets is_game_over to True when
it tries to call spawn new block, but is not able to.